requirements.txt
streamlit==1.39.0
PyMuPDF==1.24.9
openai==1.45.0
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
